requirements.txt
# =============================================================================
# SynthMT - Core Dependencies
# =============================================================================
# Python=3.11 required
# Apple Silicon M3 compatibility notes:
# If you encounter TensorFlow/Keras conflicts during installation, run:
# pip uninstall pyarrow
# This fixes compatibility issues between tf-keras and Apple Silicon optimizations.

# Core scientific computing
numpy<2.0,>=1.26
scipy~=1.16.2
scikit-learn~=1.6.1
scikit-image~=0.25.2
pandas~=2.2.3

# Image processing
pillow~=11.3.0
imageio~=2.37.0
tifffile~=2025.6.11
albumentations~=2.0.8

# Deep learning core
torch~=2.7.1
torchvision~=0.22.0
tensorflow-macos==2.16.2
tf-keras==2.16.0

# Transformers (later install pre-release for SAM3 support)
transformers~=4.56.1

# HuggingFace ecosystem
datasets~=4.5
huggingface_hub~=0.34.4

# Visualization
matplotlib~=3.10.5
seaborn~=0.13.2

# Optimization
optuna~=4.4.0

# Utilities
tqdm~=4.67.1
pyyaml~=6.0.2
python-dotenv~=1.1.1
cachetools # fix version>=5.3.0

# =============================================================================
# Model-Specific Dependencies (Optional - install as needed)
# =============================================================================

# Cellpose & CellposeSAM
cellpose~=4.0.6

# CellSAM (requires DEEPCELL_ACCESS_TOKEN in .env)
git+https://github.com/mario-koddenbrock/cellSAM.git

# StarDist (requires tensorflow)
stardist==0.9.1

# TARDIS
tardis-em==0.3.10

# LPIPS (for perceptual similarity)
lpips~=0.1.4

# =============================================================================
# Development & Testing
# =============================================================================
pytest~=8.3.5
pytest-cov~=7.0.0
tox>=4.0.0
